backend/requirements.txt
django>=4.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
requests
openai
google-generativeai
